requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ollama==0.3.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0